requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
gunicorn
django-heroku
